product/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pika==1.3.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
